snmp-service/requirements.txt
pysnmp
fastapi
uvicorn
python-dotenv